+++ gl-core/src/main/java/com/gl/service/impl/GlMesServiceImpl.java
@@ -1,6 +1,5 @@
 package com.gl.service.impl;
 
-import com.baomidou.mybatisplus.core.conditions.update.UpdateWrapper;
 import com.baomidou.mybatisplus.extension.service.impl.ServiceImpl;
 import com.gl.domain.GlAccount;
 import com.gl.domain.GlMes;
@@ -8,15 +7,9 @@ import com.gl.dto.GlMesReq;
 import com.gl.exception.BizException;
 import com.gl.service.GlMesService;
 import com.gl.mapper.GlMesMapper;
-import com.gl.service.UserService;
 import com.gl.util.UserProvider;
 import org.apache.commons.lang3.StringUtils;
-import org.springframework.security.core.GrantedAuthority;
-import org.springframework.security.core.userdetails.UserDetails;
-import org.springframework.security.core.userdetails.UserDetailsService;
 import org.springframework.stereotype.Service;
-
-import java.util.Collection;
 
 /**
 * @author Administrator
@@ -36,7 +29,7 @@ public class GlMesServiceImpl extends ServiceImpl<GlMesMapper, GlMes>
     }
 
     @Override
-    public void modifyMes(GlMesReq mesReq,Integer id) throws BizException {
+    public void modifyMes(GlMesReq mesReq,Long id) throws BizException {
         GlMes glMes = convertGlMesReq(mesReq);
         glMes.setId(id);
         baseMapper.updateById(glMes);
